extracted more code to partials
