+++ ModelFactories/ModelFactory.cs
@@ -58,39 +58,6 @@ public abstract partial class ModelFactory<T> where T : class, new()
         return list;
     }
 
-    private bool WasRecycled(PropertyInfo prop, T model)
-    {
-        if (_recycledObjects.TryGetValue(prop.PropertyType.FullName!, out var recycled))
-        {
-            prop!.SetValue(model, recycled);
-            return true;
-        }
-
-        return false;
-    }
-
-    private void ApplyProperty(T model, IPropertyDefinition propertyDefinition, bool withModel = false)
-    {
-        var reflection = model.GetType();
-        var prop = reflection.GetProperty(propertyDefinition.PropertyName);
-
-        EnsurePropExistsAndIsWritable(propertyDefinition, prop, reflection);
-
-        // Recycled values take priority over regular definitions
-        if (WasRecycled(prop!, model))
-        {
-            return;
-        }
-
-        if (withModel)
-        {
-            prop!.SetValue(model, propertyDefinition.Callback.DynamicInvoke(model));
-            return;
-        }
-
-        prop!.SetValue(model, propertyDefinition.Callback.DynamicInvoke());
-    }
-
     private void Configure()
     {
         Definition();
@@ -143,67 +110,4 @@ public abstract partial class ModelFactory<T> where T : class, new()
 
         throw new ModelFactoryException("Could not extract name from expression");
     }
-
-    #region Recycling
-
-    public ModelFactory<T> Recycle<TModel>(TModel recycledModel)
-    {
-        if (recycledModel is null)
-        {
-            throw new ModelFactoryException("Cannot recycle null");
-        }
-
-        _recycledObjects.TryAdd(typeof(TModel).FullName!, recycledModel);
-
-        return this;
-    }
-
-    internal ModelFactory<T> SetRecycledObjects(Dictionary<string, object> recycledObjects)
-    {
-        _recycledObjects = recycledObjects;
-
-        return this;
-    }
-
-    #endregion
-
-    #region Property
-
-    public ModelFactory<T> Property<TProperty>(
-        Expression<Func<T, TProperty>> propertyExpression,
-        Func<TProperty> callback
-    )
-    {
-        var propertyName = PropertyName(propertyExpression);
-        RemovePropertyKeysIfExists(propertyName);
-        _definitions.Add(propertyName, new PropertyDefinition<TProperty>(propertyName, callback));
-
-        return this;
-    }
-
-    public ModelFactory<T> Property<TProperty>(
-        Expression<Func<T, TProperty>> propertyExpression,
-        Func<T, TProperty> callback
-    )
-    {
-        var propertyName = PropertyName(propertyExpression);
-        RemovePropertyKeysIfExists(propertyName);
-        _definitionsWithModel.Add(propertyName, new PropertyDefinitionWithModel<T, TProperty>(propertyName, callback));
-
-        return this;
-    }
-
-    public ModelFactory<T> Property<TProperty>(
-        Expression<Func<T, TProperty>> propertyExpression,
-        TProperty value
-    )
-    {
-        var propertyName = PropertyName(propertyExpression);
-        RemovePropertyKeysIfExists(propertyName);
-        _definitions.Add(propertyName, new PropertyDefinition<TProperty>(propertyName, () => value));
-
-        return this;
-    }
-
-    #endregion
 }
